Split systems among updaters and renderers
