+++ internal/systems/renderers/sprite.go
@@ -1,4 +1,4 @@
-package systems
+package renderers
 
 import (
 	"reflect"
