+++ internal/systems/updaters/score_system.go
@@ -1,12 +1,12 @@
-package systems
+package updaters
 
 import (
-	"github.com/juanancid/maze-adventure/internal/queries"
 	"reflect"
 
 	"github.com/juanancid/maze-adventure/internal/components"
 	"github.com/juanancid/maze-adventure/internal/entities"
 	"github.com/juanancid/maze-adventure/internal/events"
+	"github.com/juanancid/maze-adventure/internal/queries"
 )
 
 type ScoreSystem struct {
